fix gcc warning
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@6435 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/script.cpp
@@ -715,7 +715,7 @@ void CEXEBuild::ps_addtoline(const TCHAR *str, GrowBuf &linedata, StringList &hi
             if (_T('U')==s[0] && _T('+')==s[1])
             {
               TCHAR *n=s+2;
-              unsigned long utf32=_tcstoul(n,&t,16);
+              UINT32 utf32=_tcstoul(n,&t,16);
               // We only want to accept "${U+HEXDIGITS}" and not "${U+ -HEXDIGITS }"
               if (*t || _T('-')==*n || _T('+')==*n) t=0;
               if (_T(' ')==*n || _T('\t')==*n) t=0; // TODO: _istspace()?
